Added an optional $request array to class and constructor, then migrated all $_REQUEST accessors to use $this->request.
This will allow a framework like Laravel, running with Octane + Roadrunner in a non-traditional request / response environment, to pass through the correct request data where $_REQUEST isn't set (or necessarily set for the session correctly).

The constructor will default to $_REQUEST for backwards compatibility if $request is not passed through.